App startup updates resulting in minimal logic in the custom Application class.
* Create a TimberStartupInitializer
* Move the koin logic into a KoinStartupInitializer
* Use AppStartupInitializer as the root initializer that is dependent on all other initializers
